requirements.txt
Django==3.2.13
django-environ==0.4.5
django-widget-tweaks==1.4.1
